requirements.txt
torch
transformers
datasets
accelerate
bitsandbytes
scipy
Pillow
